Fix null filenames from missing files in downloads (#1747)
In the `_all_files.zip` download, we were incorrectly including rows for
student submissions that had missing files. That is, the question
requires a file upload, but the student made an invalid submission
without a file. We were including this submission row in the SELECT,
resulting in a row with `filename = NULL` and hence a `null` file in the
zip file. To avoid this, we filter out submission rows for which the
file is missing (`filename IS NULL OR contents IS NULL`).

Closes #1744