Fix workspace reboot (#2976)
PR #2937 fixed the use of `initial.zip`, but this had the unfortunate
side-effect of breaking reboots. The problem is that we were _always_
setting `useInitialZip: true` when starting a stopped or uninitialized
workspace. In fact we only want to use the initial zip when the
workspace is uninitialized. When it's stopped, we just want to start it
again without overwriting the home directory.