Fix flapping during external grader autoscaling (#3323)
* Fix flapping during external grader autoscaling

The external grader fleet autoscaling code using multiple metrics to
estimate the desired number of graders. Most of these take a `max()`
over recent historical data and so are not subject to "flapping" (rapid
upscaling followed by immediate downscaling). However, we also use the
instantaneous number of ungraded jobs to compute the desired number of
graders, which can rapidly fluctuate. To avoid the flapping that this
can cause, this PR unconditionally latches the desired number of graders
to a `max()` over the recent past. This will avoid flapping due to
upscaling-then-downscaling. We can still have downscaling-then-upscaling
flapping, but we want that because it's caused by the need to respond to
actual load increases.

* fix typo