Add symbolic inverse trig functions (#4217)
* Add symbolic inverse trig support

PrairieLearn doesn't support the inverse trigonometric functions arcsin, arccos, and arctan. This adds support for their use in symbolic answers, which seems important as their absence renders a great deal of calculus problems typical of first year unanswerable. arcsin, arccos, and arctan were chosen for the names despite asin, acos, and atan being used by SymPy, as I think the "arc-" prefix is more common and familiar to students. If any further justification is necessary, the "arc-" prefix is also specified by ISO 80000-2.

* Add acos, asin, atan names and add to operators

Add more names for arccos, arcsin, and arctan to _Constants.functions, making them also accessible as acos, asin, and atan, as they are called in SymPy. Also add all of these to the list of operators so that their availability is obvious to frontend users.

* Add sympy.atan2

This adds the sympy.atan2 function as both arctan2 and atan2, and adds them both to the operators list.