Handle request for new variant when assessment is not available (#9803)
* Handle request for new variant when assessment is not available

If a non-broken variant exists, return the most recent one.
Otherwise, tell the user that a new variant cannot be generated.

* Formatting fixes

* Address PR comments

* Wording change

* Address nits

* Fix a flaky test

Exam questions are randomized, so we might visit the
"broken generation" question and assume that a variant
was created. Let's just remove the broken question
from the test exam.